+++ OUDS/Core/Components/Sources/Controls/CheckboxPicker/OUDSCheckboxPicker.swift
@@ -89,7 +89,7 @@ import SwiftUI
 ///
 /// ### Orange
 ///
-/// ![A checkbox picker component in light and dark mode with Orange theme](component_checkboxPicker_Orange)
+/// ![A checkbox picker component in light and dark modes with Orange theme](component_checkboxPicker_Orange)
 ///
 /// - Since: 0.14.0
 @available(iOS 15, macOS 15, visionOS 1, *)