+++ OUDS/Core/Components/Sources/Controls/RadioPicker/OUDSRadioPicker.swift
@@ -83,19 +83,19 @@ import SwiftUI
 ///
 /// ### Orange
 ///
-/// ![A radio picker component in light and dark mode with Orange theme](component_radioPicker_Orange)
+/// ![A radio picker component in light and dark modes with Orange theme](component_radioPicker_Orange)
 ///
 /// ### Orange Business Tools
 ///
-/// ![A radio picker component in light and dark mode with Orange Business Tools theme](component_radioPicker_OrangeBusinessTools)
+/// ![A radio picker component in light and dark modes with Orange Business Tools theme](component_radioPicker_OrangeBusinessTools)
 ///
 /// ### Sosh
 ///
-/// ![A radio picker component in light and dark mode with Sosh theme](component_radioPicker_Sosh)
+/// ![A radio picker component in light and dark modes with Sosh theme](component_radioPicker_Sosh)
 ///
 /// ### Wireframe
 ///
-/// ![A radio picker component in light and dark mode with Wirefraùe theme](component_radioPicker_Wireframe)
+/// ![A radio picker component in light and dark modes with Wirefraùe theme](component_radioPicker_Wireframe)
 ///
 /// - Since: 0.14.0
 @available(iOS 15, macOS 15, visionOS 1, *)