+++ OUDS/Core/Components/Sources/Layouts/Divider/OUDSHorizontalDivider.swift
@@ -33,19 +33,19 @@ import SwiftUI
 ///
 /// ### Orange
 ///
-/// ![An horizontal divider component in light and dark mode with Orange theme](component_horizontalDivider_Orange)
+/// ![An horizontal divider component in light and dark modes with Orange theme](component_horizontalDivider_Orange)
 ///
 /// ### Orange Business Tools
 ///
-/// ![A horizontal divider component in light and dark mode with Orange Business Tools theme](component_horizontalDivider_OrangeBusinessTools)
+/// ![A horizontal divider component in light and dark modes with Orange Business Tools theme](component_horizontalDivider_OrangeBusinessTools)
 ///
 /// ### Sosh
 ///
-/// ![A horizontal divider component in light and dark modeO with Sosh theme](component_horizontalDivider_Sosh)
+/// ![A horizontal divider component in light and dark modes with Sosh theme](component_horizontalDivider_Sosh)
 ///
 /// ### Wireframe
 ///
-/// ![A horizontal divider component in light and dark mode with Wireframe theme](component_horizontalDivider_Wireframe)
+/// ![A horizontal divider component in light and dark modes with Wireframe theme](component_horizontalDivider_Wireframe)
 ///
 /// - Version: 1.0.0 (Figma component design version)
 /// - Since: 0.14.0